src/test/java/no/kristiania/http_server/HttpClientTest.java
package no.kristiania.http_server;

import org.junit.jupiter.api.Test;

import java.io.IOException;

import static org.junit.jupiter.api.Assertions.*;

class HttpClientTest {
    @Test
    void shouldReadSuccessStatusCode() throws IOException {
        HttpClient httpClient = makeEchoRequest("/echo?status=200");
        assertEquals(200, httpClient.getResponseCode());
    }

    private HttpClient makeEchoRequest(String requestTarget) throws IOException {
        return new HttpClient("urlecho.appspot.com", 80, requestTarget);
    }

    @Test
    void shouldReadFailureStatusCode() throws IOException {
        HttpClient httpClient = makeEchoRequest("/echo?status=401");
        assertEquals(401, httpClient.getResponseCode());
    }

    @Test
    void shouldReadHeaders() throws IOException {
        HttpClient httpClient = makeEchoRequest("/echo?body=Kristiania");
        assertEquals("10", httpClient.getResponseHeader("Content-Length"));
    }

}